+++ code/.idea/deployment.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="PublishConfigData" serverName="HenryHuang@35.230.69.66:22 (1)">
+    <serverData>
+      <paths name="HenryHuang@35.230.69.66:22 (1)">
+        <serverdata>
+          <mappings>
+            <mapping deploy="/differential_privacy" local="$PROJECT_DIR$" web="/" />
+          </mappings>
+        </serverdata>
+      </paths>
+    </serverData>
+  </component>
+</project>
