+++ code/.idea/misc.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="JavaScriptSettings">
+    <option name="languageLevel" value="ES6" />
+  </component>
+  <component name="ProjectRootManager" version="2" project-jdk-name="Remote Python 2.7.15 (sftp://HenryHuang@35.230.69.66:22/home/HenryHuang/anaconda2/bin/python)" project-jdk-type="Python SDK" />
+</project>
