+++ code/.idea/remote-mappings.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="RemoteMappingsManager">
+    <list>
+      <list>
+        <remote-mappings server-id="python@ssh://HenryHuang@35.230.69.66:22/home/HenryHuang/anaconda2/bin/python" />
+      </list>
+    </list>
+  </component>
+</project>
