--- log.txt
@@ -1,30 +1,0 @@
-794.526021606 0.0001
-794.507768184 0.0101
-792.916525676 0.0201
-793.287903728 0.0301
-793.494504732 0.0401
-793.635651048 0.0501
-793.799881435 0.0701
-793.729374254 0.0601
-793.65286314 0.0801
-793.709937058 0.0901
-793.795012944 0.1101
-793.756524455 0.1001
-793.854783816 0.1301
-793.878418549 0.1401
-793.827312826 0.1201
-793.898957699 0.1501
-793.932885394 0.1701
-793.91696458 0.1601
-793.947304707 0.1801
-794.060602772 0.1901
-794.067631373 0.2001
-794.073938006 0.2101
-794.084778988 0.2301
-794.079625546 0.2201
-794.093753983 0.2501
-794.089469689 0.2401
-794.099243031 0.2601
-794.10291111 0.2701
-794.106200266 0.2801
-794.109243662 0.2901
