+++ report/Meeting Note 2018.07.17.txt
@@ -1,0 +1,13 @@
+Meeting Note 2018.07.17
+
+1. （Slides Page 2 问题）改变地点数时，考虑下面两种方法之一：一是相同区域，相同用户池，将粒度变得更细，二是扩大区域，扩大用户池，但选用所有用户，不进行重采样。前者更有实际意义。
+
+2. （Slides Page 3 问题）加入噪声后，展示结果时不受噪声影响的结果应当保持一致（如random choice），可通过多次实验求平均值确定。
+
+3. 需要增加其他数据集的结果，social graph采用朋友属性更强的数据集，如Facebook，之后重新实验。
+
+4. 如果最终大用户池、大志愿者数的情形仍不能有显著的效果提升，考虑将小样本数的结果作为主结果，大用户池、大志愿者数的情形作为拓展，但模型的意义有所降低。
+
+5. 仍需要调参（如每个用户上报的地点数），争取提升算法较之于随机选取的效果
+
+6. 可考虑采用多进程跑不同的数据集的实验，加快速度
