--- simu.json
@@ -1,1 +1,0 @@
-{"dir_name": "test", "dataset_num": 1, "method_num": 3, "iter_per_set": 5, "change_paras": [0], "pic_name": "SG_4_test.png", "data_src": "SG"}
